Changes: • Interface property changes and renamed file • Decided final route for the room • Updated navbar css for smaller screen
