+++ src/app/app.component.html
@@ -2,7 +2,7 @@
 <div class="flex flex-col h-screen w-screen items-center">
   <div class="w-full h-full min-w-[20rem] max-w-[80rem]">
     <!-- Show navbar -->
-    <div class="navbar-wrapper h-[4rem] w-full p-4">
+    <div class="navbar-wrapper h-[3rem] sm:h-[4rem] w-full p-4">
       <mc-navbar [user]="user()" />
     </div>
     <!-- Show Main content -->
