+++ src/app/modules/home/@interface/index.ts
@@ -1,4 +1,4 @@
-import { BasicApiResponse } from "@shared/@interface";
+import { BasicApiResponse, RoomDetails } from "@shared/@interface";
 
 export interface HomeDialog {
   isVisible: boolean;
@@ -9,21 +9,8 @@ export interface JoinRoomApiPayload {
   code: string;
 }
 
-interface JoinRoomApiResponseAttribute {
-  code: string;
-  whitePlayer: string;
-  blackPlayer: string;
-}
-
-interface CreateRoomApiResponseAttribute {
-  id: string;
-  code: string;
-  whitePlayer: string;
-  blackPlayer: string;
-  lastActivity: Date;
-  roomStatus: string;
-  gameStatus: string;
-}
+interface JoinRoomApiResponseAttribute extends RoomDetails { }
+interface CreateRoomApiResponseAttribute extends RoomDetails { }
 
 export interface JoinRoomApiResponse extends BasicApiResponse<JoinRoomApiResponseAttribute> { }
 export interface CreateRoomApiResponse extends BasicApiResponse<CreateRoomApiResponseAttribute> { }
