+++ src/app/modules/shared/@interface/index.ts
@@ -1,4 +1,4 @@
 export type * from './environment';
-export type * from './user';
+export type * from './entity';
 export type * from './ui';
 export type * from './api';
