--- src/app/modules/shared/@interface/user.ts
@@ -1,7 +1,0 @@
-export interface UserDetails {
-  username: string;
-  email: string;
-}
-
-export type UserInterface = { isLoggedIn: false; details: null; } | { isLoggedIn: true; details: UserDetails; };
-
